Clean up some mess.
Missing a return in a catch block probably meant that dataValid could have the
wrong value (i.e. be true when there were no data).

Provide a close() method on GNCDataHandler so that it can close sqliteHandle
before we lose it. This prevents sqlite getting annoyed that we didn't close
the database.

Simple aesthetic change to canReadData().

Signed-off-by: Bruce Duncan <bwduncan@gmail.com>

git-svn-id: http://gnc4a.googlecode.com/svn/trunk@76 6708f7a8-2137-5747-2049-8c0d350244a7