Fix currency detection.
The currency was never detected properly for me. My android emulator is in
American, but all of my gnucash files are in British. So the code tried to find
USD, then USD, then it bailed.

Instead, look through the list of commodities which are currencies and pick the
first one. If there are more than one, and one of the "later" currencies
has the same mnemonic as the default Locale, choose that one.

Signed-off-by: Bruce Duncan <bwduncan@gmail.com>

git-svn-id: http://gnc4a.googlecode.com/svn/trunk@87 6708f7a8-2137-5747-2049-8c0d350244a7